requirements.txt
Django==1.11.11
PyMySQL==0.8.0
djangorestframework==3.8.2
redis==2.10.6
